refactor: separated consts vs actual config
